changes based on comments #5496
