+++ src/Service/DashboardTabPanel/TabDataManager.php
@@ -10,6 +10,7 @@ use App\Entity\Territory;
 use App\Entity\User;
 use App\Repository\JobEventRepository;
 use App\Repository\PartnerRepository;
+use App\Repository\Query\Dashboard\SignalementsSansAffectationAccepteeQuery;
 use App\Repository\SignalementRepository;
 use App\Repository\SuiviRepository;
 use App\Repository\TerritoryRepository;
@@ -35,6 +36,7 @@ class TabDataManager
         private readonly PartnerRepository $partnerRepository,
         private readonly SignalementRepository $signalementRepository,
         private readonly TabCountKpiBuilder $tabCountKpiBuilder,
+        private readonly SignalementsSansAffectationAccepteeQuery $signalementsSansAffectationAccepteeQuery,
     ) {
     }
 
@@ -496,7 +498,7 @@ class TabDataManager
     {
         /** @var User $user */
         $user = $this->security->getUser();
-        $signalements = $this->signalementRepository->findSignalementsSansAffectationAcceptee(user: $user, params: $tabQueryParameters);
+        $signalements = $this->signalementsSansAffectationAccepteeQuery->findSignalements(user: $user, params: $tabQueryParameters);
         $tabDossiers = [];
         for ($i = 0; $i < \count($signalements); ++$i) {
             $signalement = $signalements[$i];
@@ -511,7 +513,7 @@ class TabDataManager
                 nbAffectations: $signalement['nbAffectations'],
             );
         }
-        $count = $this->signalementRepository->countSignalementsSansAffectationAcceptee(user: $user, params: $tabQueryParameters);
+        $count = $this->signalementsSansAffectationAccepteeQuery->countSignalements(user: $user, params: $tabQueryParameters);
 
         return new TabDossierResult($tabDossiers, $count);
     }
