+++ tests/Functional/Repository/SignalementRepositoryTest.php
@@ -449,101 +449,6 @@ class SignalementRepositoryTest extends KernelTestCase
         }
     }
 
-    public function testCountSignalementsSansAffectationAcceptee(): void
-    {
-        /** @var SignalementRepository $repo */
-        $repo = $this->entityManager->getRepository(Signalement::class);
-        /** @var UserRepository $userRepository */
-        $userRepository = $this->entityManager->getRepository(User::class);
-
-        $user = $userRepository->findOneBy(['email' => self::USER_PARTNER_TERRITORY_13]);
-        $this->assertNotNull($user);
-
-        $params = new TabQueryParameters();
-        $params->partners = [];
-        $params->mesDossiersAverifier = null;
-        $params->queryCommune = null;
-
-        $count = $repo->countSignalementsSansAffectationAcceptee($user, $params);
-
-        $this->assertIsInt($count);
-        $this->assertGreaterThanOrEqual(0, $count);
-    }
-
-    public function testFindSignalementsSansAffectationAcceptee(): void
-    {
-        /** @var SignalementRepository $repo */
-        $repo = $this->entityManager->getRepository(Signalement::class);
-        /** @var UserRepository $userRepository */
-        $userRepository = $this->entityManager->getRepository(User::class);
-
-        $user = $userRepository->findOneBy(['email' => self::USER_PARTNER_TERRITORY_13]);
-
-        $params = new TabQueryParameters();
-
-        $results = $repo->findSignalementsSansAffectationAcceptee($user, $params);
-
-        $this->assertIsArray($results);
-
-        foreach ($results as $row) {
-            $this->assertArrayHasKey('id', $row);
-            $this->assertArrayHasKey('uuid', $row);
-            $this->assertArrayHasKey('reference', $row);
-            $this->assertArrayHasKey('adresse', $row);
-            $this->assertArrayHasKey('parc', $row);
-            $this->assertArrayHasKey('nbAffectations', $row);
-            $this->assertArrayHasKey('lastAffectationAt', $row);
-            $this->assertIsNumeric($row['nbAffectations']);
-        }
-    }
-
-    public function testSignalementsSansAffectationAccepteeExclutStatutsActifs(): void
-    {
-        /** @var SignalementRepository $repo */
-        $repo = $this->entityManager->getRepository(Signalement::class);
-        /** @var UserRepository $userRepository */
-        $userRepository = $this->entityManager->getRepository(User::class);
-
-        $user = $userRepository->findOneBy(['email' => self::USER_PARTNER_TERRITORY_13]);
-
-        $params = new TabQueryParameters();
-
-        $results = $repo->findSignalementsSansAffectationAcceptee($user, $params);
-
-        foreach ($results as $row) {
-            $signalement = $this->entityManager->getRepository(Signalement::class)->find($row['id']);
-
-            foreach ($signalement->getAffectations() as $affectation) {
-                $this->assertNotContains(
-                    $affectation->getStatut(),
-                    ['EN_COURS', 'FERME'],
-                    'Un signalement retourné contient une affectation interdite'
-                );
-            }
-        }
-    }
-
-    public function testFindSignalementsSansAffectationAccepteeSortByAffectedAt(): void
-    {
-        $repo = $this->entityManager->getRepository(Signalement::class);
-        $user = $this->entityManager->getRepository(User::class)
-            ->findOneBy(['email' => self::USER_PARTNER_TERRITORY_13]);
-
-        $params = new TabQueryParameters(
-            sortBy: 'affectedAt',
-            orderBy: 'DESC'
-        );
-
-        $results = $repo->findSignalementsSansAffectationAcceptee($user, $params);
-
-        $dates = array_column($results, 'lastAffectationAt');
-
-        $sorted = $dates;
-        rsort($sorted);
-
-        $this->assertEquals($sorted, $dates);
-    }
-
     public function testFindOneForLoginBailleur(): void
     {
         /** @var SignalementRepository $signalementRepository */
